requirements.txt
discord.py==2.4.0
SQLAlchemy==2.0.32
python-dotenv==1.0.1
aiohttp==3.9.5